Deleting obsolet files, update imports routes, reorganize files and dirs
