+++ src/App.jsx
@@ -1,7 +1,7 @@
 import {Routes, Route} from 'react-router-dom'
 import { Header, Home, Footer } from './components';
 import { ProductDetail } from './components/ProductDetail/ProductDetail';
-import ProductList2 from './components/ProductList/ProductList2/ProductList';
+import ProductList from './components/ProductList/ProductList';
 import { SearchResults } from './components/SearchResults/SearchsResults';
 import { ROUTES } from './utils/constants';
 
@@ -12,7 +12,7 @@ function App() {
       <Routes>
         <Route path={ROUTES.base} element={<Home />} />
         <Route path={ROUTES.home} element={<Home />} />
-        <Route path={ROUTES.productList} element={<ProductList2 />} />
+        <Route path={ROUTES.productList} element={<ProductList />} />
         <Route path={`${ROUTES.productDetail}/:id`} element={<ProductDetail />} />
         <Route path={`${ROUTES.searchResults}`} element={<SearchResults />} />
       </Routes>
